requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
sniffio==1.3.1
spotipy==2.25.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
